organize mesos metrics collector into its own file
This commit moves the Mesos-specific collection functions and structures
into their own file, including tests. It also adds a new file "host.go"
that will be used for collecting OS-level metrics (via gopsutil).
