+++ collector/agent_test.go
@@ -18,8 +18,6 @@ package collector
 
 import (
 	"encoding/json"
-	"net/http"
-	"net/http/httptest"
 	"net/url"
 	"strconv"
 	"strings"
@@ -28,75 +26,6 @@ import (
 
 	"github.com/dcos/dcos-metrics/producers"
 	. "github.com/smartystreets/goconvey/convey"
-)
-
-var (
-	mockAgentMetrics = []byte(`
-		{
-			"system\/cpus_total": 2,
-			"system\/load_1min": 0.1,
-			"system\/load_5min": 0.11,
-			"system\/load_15min": 0.08,
-			"system\/mem_total_bytes": 4145348608,
-			"system\/mem_free_bytes": 3349942272
-		}`)
-
-	mockAgentState = []byte(`
-		{
-			"frameworks": [
-				{
-					"id": "5349f49b-68b3-4638-aab2-fc4ec845f993-0000",
-					"name": "marathon",
-					"role": "*",
-					"executors": [
-						{
-							"id": "foo.124b1048-a17a-11e6-9182-080027fb5b88",
-							"name": "Command Executor (Task: foo.124b1048-a17a-11e6-9182-080027fb5b88) (Command: sh -c 'sleep 900')",
-							"container": "e4c2f9f6-47aa-481d-a183-a21e8435bc06",
-							"labels": [
-								{
-									"key": "somekey",
-									"value": "someval"
-								}
-							],
-							"tasks": [
-								{
-									"id": "foo.124b1048-a17a-11e6-9182-080027fb5b88",
-									"name": "foo",
-									"framework_id": "5349f49b-68b3-4638-aab2-fc4ec845f993-0000",
-									"executor_id": "",
-									"slave_id": "34b46033-69c0-4663-887c-f64b526e47a6-S0",
-									"labels": [
-										{
-											"key": "somekey",
-											"value": "someval"
-										}
-									]
-								}
-							]
-						}
-					]
-				}
-			]
-		}`)
-
-	mockContainerMetrics = []byte(`
-		[
-			{
-				"container_id": "e4faacb2-f69f-4ea1-9d96-eb06fea75eef",
-				"executor_id": "foo.adf2b6f4-a171-11e6-9182-080027fb5b88",
-				"executor_name": "Command Executor (Task: foo.adf2b6f4-a171-11e6-9182-080027fb5b88) (Command: sh -c 'sleep 900')",
-				"framework_id": "5349f49b-68b3-4638-aab2-fc4ec845f993-0000",
-				"source": "foo.adf2b6f4-a171-11e6-9182-080027fb5b88",
-				"statistics": {
-					"cpus_limit": 1.1,
-					"cpus_system_time_secs": 0.31,
-					"cpus_user_time_secs": 0.22,
-					"mem_limit_bytes": 167772160,
-					"mem_total_bytes": 4476928
-				}
-			}
-		]`)
 )
 
 func TestBuildDatapoints(t *testing.T) {
@@ -144,97 +73,6 @@ func TestNewAgent(t *testing.T) {
 		Convey("Should return an Agent when given proper inputs", func() {
 			a, err := NewAgent("/bin/echo -n 1.2.3.4", 10000, 60, make(chan<- producers.MetricsMessage))
 			So(a, ShouldHaveSameTypeAs, Agent{})
-			So(err, ShouldBeNil)
-		})
-	})
-}
-
-func TestGetContainerMetrics(t *testing.T) {
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(200)
-		w.Write(mockContainerMetrics)
-	}))
-	defer ts.Close()
-
-	Convey("When fetching container metrics", t, func() {
-		port, err := extractPortFromURL(ts.URL)
-		if err != nil {
-			panic(err)
-		}
-
-		a, _ := NewAgent("/bin/echo -n 127.0.0.1", port, 60, make(chan<- producers.MetricsMessage))
-		result, err := a.getContainerMetrics()
-
-		Convey("Should return an array of 'agentContainer' without error", func() {
-			// Ensure that we're
-			//   a) unmarshaling the data correctly, and
-			//   b) that we're getting valid types for the data (string, float, int)
-			So(result[0].ContainerID, ShouldEqual, "e4faacb2-f69f-4ea1-9d96-eb06fea75eef")
-			So(result[0].Statistics.CpusLimit, ShouldEqual, 1.1)
-			So(result[0].Statistics.MemTotalBytes, ShouldEqual, 4476928)
-			So(err, ShouldBeNil)
-		})
-	})
-}
-
-func TestGetAgentMetrics(t *testing.T) {
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(200)
-		w.Write(mockAgentMetrics)
-	}))
-	defer ts.Close()
-
-	Convey("When fetching agent metrics", t, func() {
-		port, err := extractPortFromURL(ts.URL)
-		if err != nil {
-			panic(err)
-		}
-
-		Convey("Should return an 'agentMetricsSnapshot' without error", func() {
-			a, _ := NewAgent("/bin/echo -n 127.0.0.1", port, 60, make(chan<- producers.MetricsMessage))
-			result, err := a.getAgentMetrics()
-
-			So(result.CPUsTotal, ShouldEqual, 2)
-			So(result.SystemLoad5Min, ShouldEqual, 0.11)
-			So(result.MemFreeBytes, ShouldEqual, 3349942272)
-			So(err, ShouldBeNil)
-		})
-	})
-}
-
-func TestGetAgentState(t *testing.T) {
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(200)
-		w.Write(mockAgentState)
-	}))
-	defer ts.Close()
-
-	Convey("When fetching the agent state", t, func() {
-		port, err := extractPortFromURL(ts.URL)
-		if err != nil {
-			panic(err)
-		}
-
-		Convey("Should return an 'agentState' without error", func() {
-			a, _ := NewAgent("/bin/echo -n 127.0.0.1", port, 60, make(chan<- producers.MetricsMessage))
-			result, err := a.getAgentState()
-
-			// getAgentState() returns a lot of metadata required for dcos-metrics
-			// to be useful to operators. Let's ensure that we're unmarshaling
-			// and able to return *everything* we care about.
-			So(len(result.Frameworks), ShouldEqual, 1)
-			So(len(result.Frameworks[0].Executors), ShouldEqual, 1)
-			So(result.Frameworks[0].ID, ShouldEqual, "5349f49b-68b3-4638-aab2-fc4ec845f993-0000")
-			So(result.Frameworks[0].Name, ShouldEqual, "marathon")
-			So(result.Frameworks[0].Role, ShouldEqual, "*")
-			So(result.Frameworks[0].Executors[0].ID, ShouldEqual, "foo.124b1048-a17a-11e6-9182-080027fb5b88")
-			So(result.Frameworks[0].Executors[0].Name, ShouldEqual, "Command Executor (Task: foo.124b1048-a17a-11e6-9182-080027fb5b88) (Command: sh -c 'sleep 900')")
-			So(result.Frameworks[0].Executors[0].Container, ShouldEqual, "e4c2f9f6-47aa-481d-a183-a21e8435bc06")
-			So(result.Frameworks[0].Executors[0].Labels[0].Key, ShouldEqual, "somekey")
-			So(result.Frameworks[0].Executors[0].Labels[0].Value, ShouldEqual, "someval")
 			So(err, ShouldBeNil)
 		})
 	})
